fix: made get_byterange() and bytes_to_chars() safe
If a byte was passed that does not correspond to a utf-8 character
boundary but was 'in the middle of' a codepoint, the first function
would panic, the second would return a character point even though the byte
was not at the start.

This change makes both function safe to use, a new InvalidUtf8Byte error
will be returned. For get_byterange() this comes at a performance
penalty though, so the previous version was renamed to
get_byterange_unchecked() and maintained as-is for cases where the
caller is sure the bytes will be valid.